requirements.txt
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
docopt==0.6.2
Flask==3.1.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pipreqs==0.4.13
python-dotenv==1.2.1
pyudev==0.24.4
requests==2.32.5
urllib3==2.5.0
uuid==1.30
Werkzeug==3.1.3
yarg==0.1.10
